InterfaceBuilder support
Add initWithCoder method in order to support creating a
MoSegmentedControl in InterfaceBuilder by dropping a UIView and
changing its type to MoSegmentedControl. After changing its type be
sure to add a "User Defined Runtime Attribute" of "numberOfSegments" as
a Number and set its value to the number of segments you want in your
segmented control.